Simplify the validation of templates
If we can't parse a Handlebars template to extract the required data, or
if we can parse it and we find that the required data is invalid, we
treat the template as invalid.

In other words, `requiredData` has changed type from
`ErrorsOr[RequiredTemplateData.obj]` to simply `RequiredTemplateData.obj`.

Supporting the case of "it's a valid template, sort of, but its required
data is invalid" was not useful in practice and complicated the
implementation.